requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
